requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.2.0
django-filter>=23.2
drf-yasg>=1.21.5
Pillow>=9.5.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
celery>=5.3.1
redis>=4.6.0
django-storages>=1.14.2
boto3>=1.28.0  # For AWS S3
gunicorn>=21.2.0
whitenoise>=6.5.0
requests>=2.31.0  # For testing HTTP requests
